backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
spacy==3.7.2
sentence-transformers==2.2.2
numpy==1.24.3
scipy==1.11.2
scikit-learn==1.3.1
pymongo==4.5.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
pydantic==2.4.2
Werkzeug==3.0.1
requests==2.31.0
